Gitcommits_DELTA/CascadeProjects/windsurf-project/scraper-jobs/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.1.0
python-dotenv>=1.0.0
